+++ src/widgets/series-section/index.ts
@@ -1,1 +1,0 @@
-export {SeriesSection} from './series-section'